制造假数据

+++ src/main/java/com/zmsj/phoenix/model/KeyValueDto.java
@@ -12,12 +12,12 @@ public class KeyValueDto {
   @ApiModelProperty("key")
   private String key;
   @ApiModelProperty("value")
-  private int value;
+  private Double value;
 
   public KeyValueDto() {
   }
 
-  public KeyValueDto(String key, int value) {
+  public KeyValueDto(String key, Double value) {
     this.key = key;
     this.value = value;
   }
